Fix bug in mini-doc generator
There was a bug in the previous version of the mini-doc generator such
that the mini-docs contained repeated strings. This was primarily due to
the use of pointer to a list when a copy of the list should have been
done instead. In addition, it may have been that the program being used
had errors too. I have replaced the main function with a class, and this
looks like it is working now and is easier to understand too.